requirements.txt
torch==1.2.0+cpu
torchvision==0.4.0+cpu
matplotlib==3.1.1
neat-python==0.92
numpy==1.14.3
Pillow==6.2.0